requirements.txt
anyio==2.0.0b2
asyncclick==7.0.9
idna==2.10
importlib-metadata==1.7.0
python-kasa==0.4.0.dev1
sniffio==1.1.0
zipp==3.1.0
